requirement.txt
diffusers==0.31.0
pytorch==2.1.2
torchvision==0.16.2
transformers==4.46.2
numpy==1.26.3
accelerate
omegaconf
nudenet